-create user health data object -a function inside the user client to send data

+++ Frontend/composeApp/src/commonMain/kotlin/com/teamnotfound/airise/data/serializable/HealthData.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class HealthData(
     val id: String = "",
     val userDataHealthId: String = "",
-    val sleep: Int = 0,
+    val sleep: Float = 0f,
     val steps: Int = 0,
     val caloriesBurned: Int = 0,
     val avgHeartRate: Int = 0,
